app: invite message content
