+++ packages/app/src/ui-kit/organisms/InvitationContent/index.ts
@@ -1,0 +1,1 @@
+export {default as InvitationContent} from './InvitationContent';
