--- packages/app/src/ui-kit/organisms/InviteToSpaceContent/index.ts
@@ -1,1 +1,0 @@
-export {default as InviteToSpaceContent} from './InviteToSpaceContent';
