+++ packages/app/src/ui-kit/organisms/index.ts
@@ -2,7 +2,7 @@ export * from './ViusalSettingsPanel';
 export * from './LoginView';
 export * from './EventItem';
 export * from './HighFiveContent';
-export * from './InviteToSpaceContent';
+export * from './InvitationContent';
 export * from './Stage';
 export * from './OnlineUsersList';
 export * from './WorldBuilderFooter';
